Update books.components.ts
what we do here is import Components, OnInIt, books typescript from app module and our mock data of books. add our component so we can share out our selector for HTML and css.  then create our class by class BookComponent and create a method for when a book is selected, which assigns the clicked book from the template to the component's `selectedBook`, then we add another component for our mock data for books, last we add constructor and ngOnInIt()