Update book.service.ts
we are creating a service; we use this to get data from a remote server.  In our case right now we have it setup to pull some mock-data, that we have not told this service to do anything with yet but soon will.  then we also have an injectable(), that basically is tells Angular that our service will have injected dependencies, at the moment we have not defined that either.