Update app-routing.module.ts
we will start with our NgModule for two-way binding, then import our RouterModule and Routes for RouterModule makes router directives available for use in the `AppModule` our routes to define our routes for our Angular app. we then need to import BookComponents so we can define that route using a const(which basically means this will not change unless of course we update the const ) names routes that is calling property Routes which has a path set to /books from the BooksComponent.  then we import a NgModule forRoot() which configure the app at the root level and  it will then allow us to add directives like routerLink to our homepage to navigate to other pages in our angular app. Then export for allowing AppModule to access which will then export routerModule to BooksComponents.